+++ altered/__init__.py
@@ -10,6 +10,14 @@ try:
 except ImportError:
     from contextdecorator import ContextDecorator
 
+def changers(obj):
+    """
+    Chooses suitable change operations for `obj`.
+    """
+    return (hasattr(obj, '__getitem__') and
+            (dictget, dictset, dictdel) or
+            (getattr, setattr, delattr) )
+
 class state(ContextDecorator):
     """
     This combined context manager and decorator is the main API to use
@@ -22,10 +30,7 @@ class state(ContextDecorator):
         return super(state, self).__init__()
 
     def __enter__(self):
-        if hasattr(self.orig, '__getitem__'):
-            self.getter, self.setter, self.deleter = dictget, dictset, dictdel
-        else:
-            self.getter, self.setter, self.deleter = getattr, setattr, delattr
+        self.getter, self.setter, self.deleter = changers(self.orig)
         self.diff = change(self.orig, self.getter,
                            self.setter, self.deleter, **self.attrs)
         return self
@@ -40,3 +45,18 @@ class state(ContextDecorator):
             with self:
                 return f(*args, **kwds)
         return decorated
+
+def alter(obj, **changes):
+    """
+    Alternative API entry: a two step altering procedure. Same  as
+    `state`, but returns a function that will restore the changes at a
+    later point.
+    """
+    getter, setter, deleter = changers(obj)
+
+    diff = change(obj, getter, setter, deleter, **changes)
+
+    def restoration():
+        restore(obj, diff, getter, setter, deleter)
+
+    return restoration